tests/wpt/tests/encoding-detection/ja-half-width-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ja Shift_JIS late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "Shift_JIS", 'Expected Shift_JIS');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ja-half-width-late.sub.html");
</script>
